snackonai-clips/pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "snackonaiclips"
version = "1.0.0"
description = "Turn any blog post into a LinkedIn-ready short video in seconds"
authors = [{ name = "SnackOnAI", email = "hello@snackonai.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
keywords = ["ai", "video", "linkedin", "content-creation", "llm", "tts", "moviepy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "trafilatura>=1.9.0",
    "readability-lxml>=0.8.1",
    "requests>=2.31.0",
    "openai>=1.35.0",
    "moviepy>=1.0.3",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "gTTS>=2.5.0",
    "mutagen>=1.47.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
elevenlabs = ["elevenlabs>=1.1.0"]
newspaper = ["newspaper3k>=0.2.8"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
snackonaiclips = "snackonaiclips.cli:main"

[tool.pytest.ini_options]
testpaths = ["snackonaiclips/tests"]
addopts = "-v --tb=short --cov=snackonaiclips --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
